[MISC][JL] Rename folder
